ensure markdown tooltips are supported everywhere
