+++ src/utils/ui-utils.ts
@@ -6,6 +6,7 @@ import $log from '@isrd-isi-edu/chaise/src/services/logger';
 
 import { windowRef } from '@isrd-isi-edu/chaise/src/utils/window-ref';
 import { ID_NAMES } from '@isrd-isi-edu/chaise/src/utils/constants';
+import Tooltip from 'bootstrap/js/dist/tooltip';
 
 /**
  * @param   {Node=} parentContainer - the parent container. if undefined `body` will be used.
@@ -364,4 +365,34 @@ export function waitForElementToLoad(selector: string) {
       subtree: true
     });
   });
+}
+
+/**
+ * see if there's a data-chaise-tooltip in the chidren of the given element, and turn them into proper tooltips.
+ *
+ * NOTE:
+ * I'm using bootstrap.js for this feature. this has added around 30KB to our bundles. I couldn't find a way to do this
+ * directly with react-bootstrap. but there might be a way and we should investigate later
+ */
+export function createChaiseTooltips(container: Element) {
+  const tooltipTriggerList = container.querySelectorAll('[data-chaise-tooltip]');
+  if (tooltipTriggerList && tooltipTriggerList.length > 0) {
+    tooltipTriggerList.forEach((el) => {
+      const title = el.getAttribute('data-chaise-tooltip');
+      const placement = el.getAttribute('data-chaise-tooltip-placement') || 'bottom';
+      const noIcon = el.hasAttribute('data-chaise-tooltip-no-icon');
+      if (!title) return;
+      if (!noIcon) {
+        // adding space between content and the icon is how we're making sure spacing between the two is correct.
+        // should we come up with a better solution instead?
+        el.innerHTML = el.innerHTML + ' ';
+        el.classList.add('chaise-icon-for-tooltip');
+      }
+      new Tooltip(el, {
+        title,
+        // @ts-ignore ts doesn't understand that we're actually sanitizing the value.
+        placement: ['auto', 'top', 'bottom', 'left', 'right'].indexOf(placement) !== -1 ? placement : 'bottom'
+      })
+    });
+  }
 }
